refactor: Use ducks approach to simplify redux setup
