+++ src/components/Cart/CartItems.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import ProductImage from '../Products/ProductImage';
-import { FETCH_CART_START, FETCH_CART_END } from '../../reducers/cart';
+import { FETCH_CART_START, FETCH_CART_END } from '../../ducks/cart';
 
 var api = require('../../moltin.js');
