+++ src/components/Products/ProductsContainer.js
@@ -9,7 +9,7 @@ import Loading from '../global/Loading';
 import {
   FETCH_PRODUCTS_START,
   FETCH_PRODUCTS_END
-} from '../../reducers/products';
+} from '../../ducks/products';
 
 var api = require('../../moltin.js');
