+++ src/ducks/product.js
